Split resource into materialisable and other
