+++ packages/shared-src/src/models/fhir/index.js
@@ -3,6 +3,7 @@ export * from './ImagingStudy';
 export * from './Immunization';
 export * from './Patient';
 export * from './ServiceRequest';
+export * from './ServiceRequestImaging';
 
 export * from './Job';
 export * from './JobWorker';
